fix(error): Remove nesting from errors & touched
This should reduce the time needed to write to and read
from these objects. Because all the names of the inputs
are going to be unique, the errors & touched values can
all be written to a Record<string, string|boolean> with
no issues.

Even if these paths were somehow non-unique, the previous
solution of nesting the data within nested object I think
would have had an identical problem.